requirements.txt
flask==3.0.3
pypdf==4.2.0
groq
Werkzeug==3.0.3
python-dotenv==1.0.1
supabase
